scripts: fix mecab

+++ scripts/compile/mecab.fish
@@ -17,18 +17,29 @@ function pre_build
 
     $pm -- $DEPENDENCIES
 
-    cd $XDG_PROJECTS_DIR || exit
-    aria2c "$REPO_URL"
-    tar zxfv mecab-0.996.tar.gz
-    rm mecab-0.996.tar.gz
-    cd mecab-0.996/
+    if [ -d "$XDG_PROJECTS_DIR/mecab-0.996" ]
+        cd "$XDG_PROJECTS_DIR/mecab-0.996"
+    else
+        cd $XDG_PROJECTS_DIR || exit
+        if command -vq aria2c
+            aria2c "$REPO_URL"
+        else if command -vq wget
+            wget "$REPO_URL"
+        else
+            echo "install aria2c or wget"
+            exit
+        end
+        tar zxfv mecab-0.996.tar.gz
+        rm mecab-0.996.tar.gz
+        cd mecab-0.996/
+    end
 end
 
 function build
-    ./configure --build arm
+    ./configure --build aarch64-unknown-linux
     patch src/dictionary.cpp <~/repos/dotfiles/scripts/patches/mecab-src-dictionary.cpp.diff
-    make CFLAGS=-Wno-register CXXFLAGS=-Wno-register CPPFLAGS=-Wno-register
-    make check
+    make CXXFLAGS='-Wno-register -mno-outline-atomics'
+    make check || exit
     make install prefix=$PREFIX
 end
